Moved Playwright worker to `main` process
